move cert mapping checks to cache
